Code review rev 16: replace cout with LOG_ERROR, fix signed/unsigned and narrowing conversions, update stale comment

+++ src/engine/local_engine/localhand.cpp
@@ -84,8 +84,8 @@ LocalHand::LocalHand(boost::shared_ptr<EngineFactory> f, GuiInterface *g, boost:
 	for(it=activePlayerList->begin(); it!=activePlayerList->end(); ++it, k++) {
 
 		// Bounds check: ensure card array indices are in range.
-		// k is bounded by activePlayerList->size() (max MAX_NUMBER_OF_PLAYERS=10),
-		// so max card index is 2*(size-1)+1+5 = 24, well within NumCards (52).
+		// k is bounded by activePlayerList->size() (max MAX_NUMBER_OF_PLAYERS),
+		// so max card index is 2*(size-1)+1+5, well within NumCards (52).
 		if (k >= static_cast<int>(activePlayerList->size()) || 2*k+1+5 >= NumCards) {
 			throw LocalException(__FILE__, __LINE__, ERR_SEAT_NOT_FOUND);
 		}
@@ -435,7 +435,7 @@ void LocalHand::assignButtons()
 		throw LocalException(__FILE__, __LINE__, ERR_SEAT_NOT_FOUND);
 	}
 
-	for(i=0; i<seatsList->size(); i++) {
+	for(i=0; i<static_cast<int>(seatsList->size()); i++) {
 
 		++dealerPositionIt;
 		if(dealerPositionIt == seatsList->end()) dealerPositionIt = seatsList->begin();
